fix: algunos errores y warnings varios

+++ src/Escritura/Escritura.cc
@@ -38,7 +38,7 @@ void Escritura::VolcarEnFichero(ofstream& file) {
 
 ostream& operator<<(ostream& os, const Escritura& escritura) {
     os << "Escritura: ";
-    for (int i = 0; i < escritura.cinta_escritura.size(); i++) {
+    for (long unsigned int i = 0; i < escritura.cinta_escritura.size(); i++) {
       os << escritura.cinta_escritura[i] << " ";
     }
     os << endl;
